requirements.txt
fitz==0.0.1.dev2
ghostscript==0.6
Image==1.5.33
numpy==1.17.1
pandas==1.0.3
pdf2image==1.13.1
Pillow==8.4.0
PyPDF2==1.26.0
pytest==6.2.5
PythonMagick==0.5
setuptools==41.2.0
utils==1.0.1
Wand==0.6.1
black==22.1.0
flake8==4.0.1